+++ pol-core/pol/uworld.h
@@ -161,7 +161,7 @@ namespace
 inline CoordsArea::CoordsArea( const Pos2d& p, const Realms::Realm* posrealm, unsigned range )
 {
   realm = posrealm;
-  if ( range > std::numeric_limits<s16>::max() )
+  if ( range > static_cast<u32>( std::numeric_limits<s16>::max() ) )
     range = std::numeric_limits<s16>::max();
   Vec2d r( static_cast<s16>( range ), static_cast<s16>( range ) );
   area = Area2d( p - r, p + r, realm );
@@ -170,7 +170,7 @@ inline CoordsArea::CoordsArea( const Pos2d& p, const Realms::Realm* posrealm, un
 inline CoordsArea::CoordsArea( const Pos4d& p, unsigned range )
 {
   realm = p.realm();
-  if ( range > std::numeric_limits<s16>::max() )
+  if ( range > static_cast<u32>( std::numeric_limits<s16>::max() ) )
     range = std::numeric_limits<s16>::max();
   Vec2d r( static_cast<s16>( range ), static_cast<s16>( range ) );
   area = Area2d( ( p - r ).xy(), ( p + r ).xy(), realm );
